+++ paillier-encryption/src/main/java/org/enquery/encryptedquery/encryption/paillier/PaillierKeyPair.java
@@ -55,3 +55,4 @@ class PaillierKeyPair {
 		return priv;
 	}
 }
+